pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "infra-clarity"
version = "0.1.0"
description = "A CLI tool for identifying cloud misconfigurations and cost optimizations"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Infra Clarity Team"}
]
dependencies = [
    "boto3>=1.28.0",
    "python-hcl2>=4.3.2",
    "typer>=0.9.0",
    "rich>=13.4.2",
    "pydantic>=2.0.0",
    "tabulate>=0.9.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pylint>=3.0.0"
]

dashboard = [
    "streamlit>=1.25.0",
    "plotly>=5.15.0",
    "pandas>=2.0.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["infra_clarity*"]

[project.scripts]
infra-clarity = "infra_clarity.cli:app"
